Fix course cache bug: changes in included configuration templates
The course configuration cache checks if the configurations of
the requested exercise have changed in the file system by
comparing the modification timestamp of the exercise configuration
file (often named config.yaml) to the timestamp stored in the cache.
However, some settings may be changed without modifying
the config.yaml file itself. This commit detects the changes in
the included base configuration template by checking its file
modification timestamp too. The base configuration template
is defined with the "include" field (if it is used).

The latest modification timestamp of the configuration files is
stored in the cache. That timestamp is also used outside
the config.py module, thus it must be the latest modification
timestamp of the exercise.

Fixes #105