add an intermediary build directory
Store built courses in an intermediary directory after building
but before publishing. Once A+ wants to update the course,
publish the course in the built directory and tell graders to do
the same.

This moves the configuring of the graders to the build stage and makes
the actual publishing of the changes very fast.