add absolute symlink check to built course files
Absolute symlinks are copied as-is: this would cause the course to fail
or have hard to debug problems when the file is deleted or changed in the
build directory. There shouldn't be any absolute symlinks anyway: the
courses are built inside a container with different location for
the course diretory.