Fix invalid CourseUpdate.status model field definition
The model field definition used the UpdateStatus enum incorrectly.
The status choices were strings that started with "UpdateStatus.",
e.g., "UpdateStatus.PENDING". The prefix "UpdateStatus" is not
supposed to be saved to the database. It even breaches the max_length
constraint of the CharField, but since SQLite ignores max_length,
this does not cause issues when testing on SQLite.