Set the default logging level to DEBUG for the course build
The course build log should include all the log output from the build.
It helps course staff investigate issues in the build.

The builder.builder module sets the DEBUG level in the top of the module
anyway, but it doesn't apply in all cases. At least, when running
the gitmanager locally in the Huey immediate mode, the build log output
used the level set in settings.py.