Fix parsing dates with time without quotes
Pydantic coerces the input to the Union type in the order
the types are defined in the Union. For example, the input
`2022-03-05 10:44:00` loses the time completely if it is
parsed as `date` first before trying `datetime`. That occurs
with `Union[date, datetime]`. If the input is defined as
a string with quotes `'2022-03-05 10:44:00'`, then it works
with the old version that coerces to `date` before `datetime`.

* https://pydantic-docs.helpmanual.io/usage/types/#unions
* https://pydantic-docs.helpmanual.io/usage/types/#datetime-types