Add course source directory for local testing without git
If the git origin is empty in the course settings, then the course
source files are copied from the new course source directory
(if the local source directory is enabled in the settings).
Course directories may be mounted to the source directory in local
testing so that changes in the course may be built without any
git setup.

It used to be (and still is) possible to leave the git origin empty
and mount course source directories directly to the build directory,
but that setup is prone to file permission issues. The build fails
if the build process in the container does not have write access
to the build directory.

In the run-gitmanager container image, remove COURSES_PATH
`/srv/courseshuey` since it is not useful any longer.
Unset git origin and leave it empty in the database setup
for the default course.