Keep the .git directory in the local course build testing setup
Since `git clean` is run at the end of the course build nowadays,
it caused a crash in the local testing setup when
`settings.LOCAL_COURSE_SOURCE_PATH` is used.
It crashed because the course directory was missing the `.git`
directory and thus, it was not a git repository.
The build error was the following:

```
stdout: fatal: not a git repository (or any parent up to mount point /tmp)
Stopping at filesystem boundary (GIT_DISCOVERY_ACROSS_FILESYSTEM not set)
```

Fix the crash by including the `.git` directory when
the course source directory is copied to the build directory.
The `.git` directory used to be excluded simply because the copy
operation is a bit faster without it.